examples/launchmybakery/requirements.txt
google-adk==1.28.0
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
db-dtypes
